requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
openai==1.68.2
pyyaml==6.0.1
python-dotenv==1.0.0
httpx==0.25.0
